refactor: created folders for each class
