+++ server/src/address/addressDto.ts
@@ -1,0 +1,12 @@
+import { User } from '../user/userDto';
+
+export interface Address {
+    id: number;
+    rua: string;
+    numero: number;
+    cep: string;
+    complemento?: string; // Optional
+    cidade: string; // Default: Recife
+    estado: string; // Default: Pernambuco
+    user: User;
+}
