+++ server/src/school/schoolDto.ts
@@ -1,0 +1,7 @@
+export interface School {
+    id: number;
+    cnpj: string;
+    numero_estudantes: number;
+    pontos_acumulados: number;
+    //acaoEscola: AcaoEscola;
+}
